// angular-app/src/app/core/models/program.model.ts
export interface Instruction {
  action: ProgAction;
  label: string | null;
  value: string | null;
}